Commit Summary 041525
Updated the basic guide and the lighting device guide with better-formatted versions. Added the speaker and bumper guides. Replaced the sta file with a zip file containing imperial and metric templates and examples. Revised the revision history file. Updated the description of all of these on the
main Content Development page.